move bilby sampling to sample project
