--- projects/sandbox/bilby/pyproject.toml
@@ -1,24 +1,0 @@
-[tool.poetry]
-name = "bilby-analysis"
-version = "0.1.0"
-description = "Produce posterior samples with bilby to compare with our network"
-authors = ["ethan.marx <ethan.marx@ligo.org>"]
-license = "MIT"
-
-[tool.poetry.dependencies]
-python = ">=3.9,<3.11"
-typeo = {git = "https://github.com/ML4GW/typeo.git"}
-"mlpe.logging" = {path = "../../../libs/logging", develop = true}
-"mlpe.injection" = {path = "../../../libs/injection", develop = true} # needed for loading sg waveform
-bilby = "^2.0.0"
-bilby_pipe = "^1.0.8"
-
-[tool.poetry.scripts]
-analyze-with-bilby = "bilby_analysis:main"
-
-[tool.poetry.group.dev.dependencies]
-ipykernel = "^6.21.3"
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
